Move browser launch helpers out of running session
