+++ tests/Aspire.Hosting.Tests/BrowserLogsSessionManagerTests.cs
@@ -17,29 +17,6 @@ namespace Aspire.Hosting.Tests;
 public class BrowserLogsSessionManagerTests
 {
     [Fact]
-    public void TryParseBrowserDebugEndpoint_ReturnsBrowserWebSocketUri()
-    {
-        var endpoint = BrowserLogsDebugEndpointParser.TryParseBrowserDebugEndpoint("""
-            51943
-            /devtools/browser/4c8404fb-06f8-45f0-9d89-112233445566
-            """);
-
-        Assert.NotNull(endpoint);
-        Assert.Equal("ws://127.0.0.1:51943/devtools/browser/4c8404fb-06f8-45f0-9d89-112233445566", endpoint.AbsoluteUri);
-    }
-
-    [Theory]
-    [InlineData("")]
-    [InlineData("not-a-port")]
-    [InlineData("51943")]
-    public void TryParseBrowserDebugEndpoint_ReturnsNullForInvalidMetadata(string metadata)
-    {
-        var endpoint = BrowserLogsDebugEndpointParser.TryParseBrowserDebugEndpoint(metadata);
-
-        Assert.Null(endpoint);
-    }
-
-    [Fact]
     public void BrowserHostIdentity_NormalizesTrailingDirectorySeparators()
     {
         WithTempUserDataDirectory(userDataDirectory =>
